use double not float
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@59942 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/renderer-desktop/src/main/java/com/himamis/retex/renderer/desktop/geom/Point2DD.java
@@ -45,7 +45,7 @@ package com.himamis.retex.renderer.desktop.geom;
 
 import java.awt.geom.Point2D;
 
-public class Point2DD extends Point2D.Float implements com.himamis.retex.renderer.share.platform.geom.Point2D {
+public class Point2DD extends Point2D.Double implements com.himamis.retex.renderer.share.platform.geom.Point2D {
 
 	private static final long serialVersionUID = 1L;
 
@@ -55,12 +55,12 @@ public class Point2DD extends Point2D.Float implements com.himamis.retex.rendere
 	
 	@Override
 	public void setX(double x) {
-		this.x = (float) x;
+		this.x = x;
 	}
 
 	@Override
 	public void setY(double y) {
-		this.y = (float) y;
+		this.y = y;
 	}
 
 }
